+++ packages/proxy/src/upstream/index.ts
@@ -1,3 +1,4 @@
+// eslint-disable-next-line @typescript-eslint/no-deprecated -- deliberate compat re-export of the deprecated alias
 export { UpstreamForwarder } from './forwarder.js'
 export type { UpstreamForwarderOptions } from './forwarder.js'
 export { SseUpstreamForwarder } from './sse-forwarder.js'